Store loaders discard a corrupt payload instead of raising out of setup (#238)
A non-mapping .storage file was crashing six fire-and-forget loaders on stored.get; thermal_learning never self-healed. The parse now sits inside the existing try (dict() copy / isinstance mapping check) and from_dict catches OverflowError/KeyError. Recorded coordinator_loc 10365→10363, max_class_loc 10365→10363, cut_dhw 195→194 — the normalize restructure paid for the guards.

+++ custom_components/heatpump_optimizer/dhw_draws.py
@@ -124,7 +124,7 @@ class DrawStats:
                     stats.reservoirs[str(label)] = [
                         float(e)
                         for e in events[-MAX_EVENTS_PER_WINDOW:]
-                        if isinstance(e, (int, float)) and np.isfinite(e)
+                        if isinstance(e, (int, float)) and np.isfinite(np.float64(e))
                     ]
         stats._open_label = str(data.get("open_label", ""))
         stats._open_date = str(data.get("open_date", ""))
